pyproject.toml
[tool.poetry]
name = "egovernance_project"
version = "0.1.0"
description = ""
authors = ["ShyamSundhar1411 <65313257+ShyamSundhar1411@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
Pillow = "^9.1.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
django-autoslug = "^1.9.8"
social-auth-app-django = "^5.0.0"
django-crispy-forms = "^1.14.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^6.1.0"}
python-social-auth = {extras = ["django"], version = "^0.3.6"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
